fix: Move logic from player file viewer to player view
