+++ ui/package.json
@@ -11,7 +11,8 @@
     "build-only": "vite build",
     "type-check": "vue-tsc --build",
     "lint": "eslint . --fix",
-    "format": "prettier --write src/"
+    "format": "prettier --write src/",
+    "build-dev": "vite build -h"
   },
   "dependencies": {
     "@mdi/js": "^7.4.47",
